restructure repo, containerize services
