--- go.mod
@@ -1,8 +1,0 @@
-module go-rec-sys
-
-go 1.22.1
-
-require (
-	github.com/lib/pq v1.10.9
-	golang.org/x/net v0.26.0
-)
