+++ go.work
@@ -1,0 +1,8 @@
+go 1.22.1
+
+use (
+	./libs
+	./services/crawler
+	./services/recommendation_system
+	./services/server
+)
