+++ libs/database/query/tournament.sql
