+++ libs/go.mod
@@ -1,0 +1,5 @@
+module github.com/Khiem17204/go_recommendation_system/libs
+
+go 1.22.1
+
+require github.com/lib/pq v1.10.9
