+++ services/crawler/card_helper.go
@@ -4,10 +4,11 @@ import (
 	"database/sql"
 	"encoding/json"
 	"fmt"
-	db "go-rec-sys/db/sqlc"
-	utils "go-rec-sys/libs/utils/class"
-	database "go-rec-sys/libs/utils/database"
 	"net/http"
+
+	db "github.com/Khiem17204/go_recommendation_system/libs/database/sqlc"
+	utils "github.com/Khiem17204/go_recommendation_system/libs/utils/class"
+	database "github.com/Khiem17204/go_recommendation_system/libs/utils/database"
 )
 
 type cardHelper struct {
