+++ services/crawler/processDeck.go
@@ -3,12 +3,13 @@ package main
 import (
 	"encoding/json"
 	"fmt"
-	db "go-rec-sys/db/sqlc"
-	utils "go-rec-sys/libs/utils/class"
-	database "go-rec-sys/libs/utils/database"
 	"io"
 	"net/http"
 	"strconv"
+
+	db "github.com/Khiem17204/go_recommendation_system/libs/database/sqlc"
+	utils "github.com/Khiem17204/go_recommendation_system/libs/utils/class"
+	database "github.com/Khiem17204/go_recommendation_system/libs/utils/database"
 )
 
 // https://ygoprodeck.com/api/decks/getDeckInfo.php?deckId=506208
