+++ services/server/go.mod
@@ -1,0 +1,3 @@
+module github.com/Khiem17204/go_recommendation_system/services/server
+
+go 1.22.1
